requirements.txt
# Core
Django==5.2
djangorestframework==3.16.0
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2

# Database
psycopg2-binary==2.9.9
psycopg==3.2.7

# Authentication
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0

# API Documentation
drf-spectacular==0.28.0
uritemplate==4.1.1
PyYAML==6.0.2

# Development Tools
django-debug-toolbar==5.2.0
django-extensions==4.1
python-dotenv==1.1.0
dotenv==0.9.9

# Testing
pytest==8.3.5
attrs==25.3.0
iniconfig==2.1.0
pluggy==1.5.0

# Code Quality
black==25.1.0
flake8==7.2.0
pycodestyle==2.13.0
pyflakes==3.3.2
mccabe==0.7.0
mypy_extensions==1.1.0

# Utilities
click==8.1.8
colorama==0.4.6
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
referencing==0.36.2
rpds-py==0.24.0